chore(mme): Break up mme_procedure test into multiple files
