+++ lte/gateway/c/core/oai/test/mock_tasks/mock_tasks.h
@@ -10,6 +10,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#pragma once
+
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
